+++ core/src/main/scala/com/jpro/web/Util.scala
@@ -3,7 +3,9 @@ package com.jpro.web
 import com.jpro.webapi.{HTMLView, WebAPI}
 import simplefx.core._
 import simplefx.all._
+import simplefx.experimental._
 import simplefx.util.ReflectionUtil._
+import simplefx.util.Predef._
 import java.net.URLEncoder
 
 import com.jpro.web.sessionmanager.SessionManager
@@ -37,13 +39,7 @@ object Util {
   }
   def setLinkInternal(node: Node, url: String, text: Option[String] = None, children: ObservableList[Node] = null) = {
     node.cursor = javafx.scene.Cursor.HAND
-    if(!WebAPI.isBrowser) {
-      node.onMouseClicked --> { e =>
-        if(e.isStillSincePress) Util.getSessionManager(node).goto(url)
-      }
-    } else {
-      setLinkSimple(url, text, true)(node, children)
-    }
+    setLinkSimple(url, text, true)(node, children)
   }
   def setLinkExternal(node: Node, url: String, text: Option[String] = None, children: ObservableList[Node] = null) = {
     node.cursor = javafx.scene.Cursor.HAND
@@ -58,48 +54,87 @@ object Util {
     SessionManagerContext.getContext(node).goForward()
   }
   def gotoPage(node: Node, url: String) = {
-    Util.getSessionManager(node).goto(url)
+    Util.getSessionManager(node).gotoURL(url)
   }
   def getCurrentPage(node: Node): String = {
     Util.getSessionManager(node).url
   }
   def refresh(node: Node): Unit = {
     val man = Util.getSessionManager(node)
-    man.goto(man.url)
+    man.gotoURL(man.url)
   }
 
-  private def setLinkSimple(url: String, text: Option[String], pushState: Boolean)(theNode: Node, children: ObservableList[Node] = null) = if(WebAPI.isBrowser) onceWhen(theNode.parent != null) --> {
-    assert(children != null || theNode.parent.isInstanceOf[Region], "The parent at setLink has to be a Pane")
-    //val parent = theNode.parent.asInstanceOf[Region]
+  @extension
+  class ExtendNodeWithLink(node: Node) {
+    var children: ObservableList[Node] = null
+
+    def setNewLink(link: String, text: Option[String], pushState: Boolean, children: ObservableList[Node]): Unit = {
+      this.children = children
+      this.pushState = pushState
+      this.link = link
+      this.text = text
+    }
+
+    @Bind var pushState: Boolean = false
+    @Bind var link: String = ""
+    @Bind var text: Option[String] = None
+
     val id = "linkid_"+random[Int].abs
-    val htmlNode = new HTMLView { htmlNode =>
-      layoutXY    <-- theNode.bipXY
-      theNode.bipWH --> { x =>
+    @Bind val htmlNode = new HTMLView { htmlNode =>
+      layoutXY    <-- node.bipXY
+      node.bipWH --> { x =>
         this.resize(x._1,x._2)
       }
-      val script = if(pushState) {
+      def script = if(pushState) {
         s"""<script>var x = document.getElementById("${id}");
            | x.addEventListener("click", function(event) {
            |   if(!event.shiftKey && !event.metaKey) {
-           |     jpro.jproGotoURL(\"${url.replace("\"","\\\"")}\"); event.preventDefault();
+           |     jpro.jproGotoURL(\"${link.replace("\"","\\\"")}\"); event.preventDefault();
            |   }
            | });</script>
         """.stripMargin
       } else ""
 
-      val styleAnchorText = if(text.isEmpty) "" else "line-height: 0; font-size: 0; color: transparent; "
-      setContent(
-        s"""<a id="$id" href="${url.replace(" ","%20").replace("\"","&quot;")}" style="$styleAnchorText display: block; width: 100%; height: 100%;">${text.getOrElse("")}</a>
-
+      def styleAnchorText = if(text.isEmpty) "" else "line-height: 0; font-size: 0; color: transparent; "
+      @Bind var content = contentProperty.toBindable
+      content <-- {
+        s"""<a id="$id" href="${link.replace(" ","%20").replace("\"","&quot;")}" style="$styleAnchorText display: block; width: 100%; height: 100%;">${text.getOrElse("")}</a>
+           |$script
          """.stripMargin
-      )
+      }
       hover --> { x =>
-        theNode.useReflection.setHover(x)
+        node.useReflection.setHover(x)
       }
       setManaged(false)
     }
+    if(!WebAPI.isBrowser) {
+      node.onMouseClicked --> { e =>
+        if(e.isStillSincePress) Util.getSessionManager(node).gotoURL(link)
+      }
+    } else {
+      var currentParent: ObservableList[Node] = null
+      def cleanCurrentParent(): Unit = {
+        if(currentParent != null) {
+          if(currentParent.contains(htmlNode)) {
+            currentParent.remove(htmlNode)
+          }
+          currentParent = null
+        }
+      }
+      when(node.parent != null && link != null) --> {
+        assert(children != null || node.parent.isInstanceOf[Region] || node.parent.isInstanceOf[Group], s"The parent at setLink has to be a Pane but was ${node.parent}")
+        cleanCurrentParent()
+        currentParent = if(node.parent.isInstanceOf[Region]) node.parent.asInstanceOf[Pane].getChildren
+          else if(node.parent.isInstanceOf[Group])  node.parent.asInstanceOf[Group].getChildren
+          else children
+        currentParent.add(htmlNode)
+        onDispose(cleanCurrentParent())
+      }
+    }
+  }
 
-    val theChildren = if(children == null) theNode.parent.asInstanceOf[Pane].getChildren else children
-    theChildren.add(htmlNode)
+
+  private def setLinkSimple(url: String, text: Option[String], pushState: Boolean)(theNode: Node, children: ObservableList[Node] = null) = {
+    theNode.setNewLink(url,text,pushState,children)
   }
 }