+++ example/src/main/scala/example/TestWebApplication.scala
@@ -37,7 +37,7 @@ class MyApp(stage: Stage) extends WebApp(stage) {
  // addTransition{ case (view,view2,false) => PageTransition.MoveUp }
 }
 
-class Header(var sessionManager: SessionManager) extends HBox {
+class Header(view: View, sessionManager: SessionManager) extends HBox {
   padding = Insets(10)
   spacing = 10
   class HeaderLink(str: String, url: String) extends Label (str) {
@@ -60,7 +60,17 @@ class Header(var sessionManager: SessionManager) extends HBox {
   this <++ new HeaderLink("collect"    , "/?page=collect" )
   this <++ new HeaderLink("jmemorybuddy"    , "/?page=jmemorybuddy" )
   this <++ new HeaderLink("No Link" , "" )
+  this <++ new HeaderLink("ManyLinks" , "" ) {
+    setLink(this, "/?1", Some("/?1"))
+    setLink(this, "/?2", Some("/?2"))
+    setLink(this, "/?3", Some("/?3"))
+    setLink(this, "/?4", Some("/?4"))
+    setLink(this, "/?5", Some("/?5"))
+    setLink(this, "/?6", Some("/?6"))
+    setLink(this, "/?7", Some("/?7"))
+  }
 
+  this <++ new Label(view.url)
 
   this <++ new Button("Backward") {
     disable <-- (!WebAPI.isBrowser && sessionManager.historyBackward.isEmpty)
@@ -87,19 +97,24 @@ class Footer(sessionManager: SessionManager) extends HBox {
   }
 }
 
-trait Page extends View {
+trait Page extends View { view =>
   override lazy val realContent = {
     new VBox {
   // Cousing leak? style = "-fx-background-color: white;"
     //  transform = Scale(1.3,1.3)
       spacing = 10
-      this <++ new Header(sessionManager)
+      this <++ new Header(view, sessionManager)
       val theContent = content
       javafx.scene.layout.VBox.setVgrow(theContent,Priority.ALWAYS)
       this <++ theContent
       this <++ new Footer(sessionManager)
       //this <++ new Header(sessionManager)
     }
+  }
+
+  override def handleURL(x: String): Boolean = {
+    println("handleURL called: " + x);
+    return false;
   }
 }
 
@@ -286,9 +301,7 @@ object TestWebApplication extends App
   if(WebAPI.isBrowser) {
     root = app
   } else {
-    scene = new Scene(new ScrollPane(app) {
-      fitToWidth = true
-    }, 1400,800)
+    scene = new Scene(app, 1400,800)
   }
   app.start(SessionManager.getDefault(app,stage))
 }